app/requirements.txt
Django==5.2.8

# Django REST
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.29.0

# Localization / Admin
django-modeltranslation==0.19.12
django-unfold==0.85.0
django-countries==8.2.0
django-simple-history==3.11.0

# Utils
django-environ==0.11.2
django-phonenumber-field[phonenumbers]==8.0.0
django-import-export==4.3.7
django-ckeditor-5==0.2.18

# Debug / Monitoring
django-silk==5.3.2
django-debug-toolbar==5.2.0
sentry-sdk[django]==2.26.1

# Cache / Redis
django-redis==5.4.0
redis==5.0.4

# Files / Docs
Pillow==11.2.1
qrcode==8.2
reportlab==4.4.4
openpyxl==3.1.5
python-docx==1.2.0
html-for-docx==1.1.1

# Server / DB / HTTP
gunicorn==23.0.0
httpx==0.28.1
psycopg[c]==3.2.9
